+++ FakeServer/Controllers/DynamicController.cs
@@ -79,13 +79,8 @@ namespace FakeServer.Controllers
 
             if (itemType == JsonFlatFileDataStore.ValueType.Item)
                 return GetSingleItem(collectionId);
-            else
-            {
-                if (!QueryHelper.IsQueryValid(Request.Query))
-                    return BadRequest();
-
-                return GetCollectionItem(collectionId, skip, take);
-            }
+            
+            return GetCollectionItem(collectionId, skip, take);
         }
 
         private IActionResult GetSingleItem(string itemId)
@@ -101,16 +96,15 @@ namespace FakeServer.Controllers
 
         private IActionResult GetCollectionItem(string collectionId, int skip, int take)
         {
-            var collection = _ds.GetCollection(collectionId);
-
-            // Collection can actually just be empty, but in this case we handle it as it is not found
-            if (!collection.AsQueryable().Any())
-                return NotFound();
-
+            if (!QueryHelper.IsQueryValid(Request.Query))
+                return BadRequest();
+            
             var options = QueryHelper.GetQueryOptions(Request.Query, skip, take);
 
             if (!options.Validate())
                 return BadRequest();
+
+            var collection = _ds.GetCollection(collectionId);
 
             var datas = options.IsTextSearch ? collection.Find(Request.Query["q"]) : collection.AsQueryable();